Redesign settings command
